Move get_vapoursynth_version, get_vapoursynth_api_version
